sprint5-6 исправил типы

+++ internal/app/handlers/balance/get.go
@@ -8,7 +8,7 @@ import (
 )
 
 type UserBalance struct {
-	Current   float64 `json:"current"`
+	Current   float32 `json:"current"`
 	Withdrawn int     `json:"withdrawn"`
 }
