+++ internal/app/storage/models/user_balance.go
@@ -2,6 +2,6 @@ package models
 
 type UserBalance struct {
 	ID        int     `json:"id"`
-	Balance   float64 `json:"balance"`
+	Balance   float32 `json:"balance"`
 	Withdrawn int     `json:"withdrawn"`
 }
